responsivity...

+++ profiles/views.py
@@ -10,10 +10,13 @@ from .serializers import ProfileSerializer, RelationshipSerializer, UserSerializ
 from .models import *
 
 # Create your views here.
+
+
 @api_view(['GET'])
 def get_logged_user(request):
     serializer = UserSerializer(request.user)
     return Response(serializer.data)
+
 
 @api_view(['GET'])
 def get_profile(request, slug):
@@ -24,6 +27,7 @@ def get_profile(request, slug):
     else:
         return Response({'bool': 'false'})
 
+
 @api_view(['GET'])
 def get_profile_by_email(request, email):
     if Profile.objects.filter(email=email, user__is_active=True).exists():
@@ -33,17 +37,22 @@ def get_profile_by_email(request, email):
     else:
         return Response({'bool': 'false'})
 
+
 @api_view(['GET'])
 def filter_profiles(request, query):
     profile = Profile.objects.get(user=request.user)
     profiles = []
     for p in Profile.objects.all().exclude(user=profile.user):
-        if query.lower() not in p.user.username: continue
-        if profile.user in p.blocked_users.all(): continue
-        if p.user in profile.blocked_users.all(): continue
+        if query.lower() not in p.user.username:
+            continue
+        if profile.user in p.blocked_users.all():
+            continue
+        if p.user in profile.blocked_users.all():
+            continue
         profiles.append(p)
     serializer = ProfileSerializer(profiles, many=True)
     return Response(serializer.data)
+
 
 def get_profile_list(profile):
     profiles = []
@@ -53,11 +62,16 @@ def get_profile_list(profile):
             if p in profiles:
                 shared_interests_quantity[profiles.index(p)] += 1
                 continue
-            if p.user in profile.friends.all(): continue
-            if p.user in profile.blocked_users.all(): continue
-            if profile.user in p.blocked_users.all(): continue
-            if p in [i.receiver for i in Relationship.objects.invitations_sent(profile)]: continue
-            if p in [i.sender for i in Relationship.objects.invitations_received(profile)]: continue
+            if p.user in profile.friends.all():
+                continue
+            if p.user in profile.blocked_users.all():
+                continue
+            if profile.user in p.blocked_users.all():
+                continue
+            if p in [i.receiver for i in Relationship.objects.invitations_sent(profile)]:
+                continue
+            if p in [i.sender for i in Relationship.objects.invitations_received(profile)]:
+                continue
             profiles.append(p)
             shared_interests_quantity.append(1)
     if len(profiles):
@@ -68,12 +82,16 @@ def get_profile_list(profile):
         profiles.reverse()
     else:
         for p in Profile.objects.all().exclude(user=profile.user):
-            if p.user in profile.friends.all(): continue
-            if p in [i.receiver for i in Relationship.objects.invitations_sent(profile)]: continue
-            if p in [i.sender for i in Relationship.objects.invitations_received(profile)]: continue
+            if p.user in profile.friends.all():
+                continue
+            if p in [i.receiver for i in Relationship.objects.invitations_sent(profile)]:
+                continue
+            if p in [i.sender for i in Relationship.objects.invitations_received(profile)]:
+                continue
             profiles.append(p)
         random.shuffle(profiles)
     return profiles
+
 
 @api_view(['GET'])
 def profile_list_view(request):
@@ -82,22 +100,26 @@ def profile_list_view(request):
     serializer = ProfileSerializer(profiles[:50], many=True)
     return Response(serializer.data)
 
+
 @api_view(['GET'])
 def interest_profile_list(request, interest):
     profile = Profile.objects.get(user=request.user)
     profiles = []
     for p in Profile.objects.filter(interests__title=interest.lower()):
-        if p == profile: continue
+        if p == profile:
+            continue
         profiles.append(p)
     random.shuffle(profiles)
     serializer = ProfileSerializer(profiles[:50], many=True)
     return Response(serializer.data)
+
 
 @api_view(['GET'])
 def my_profile(request):
     profile = Profile.objects.get(user=request.user)
     serializer = ProfileSerializer(profile)
     return Response(serializer.data)
+
 
 @api_view(['GET'])
 def friends_profiles(request, slug):
@@ -107,12 +129,15 @@ def friends_profiles(request, slug):
     serializer = ProfileSerializer(friends, many=True)
     return Response(serializer.data)
 
+
 @api_view(['GET'])
 def blocked_profiles(request):
     profile = Profile.objects.get(user=request.user)
-    profiles = [blocked_user.profile for blocked_user in profile.blocked_users.all()]
+    profiles = [
+        blocked_user.profile for blocked_user in profile.blocked_users.all()]
     serializer = ProfileSerializer(profiles, many=True)
     return Response(serializer.data)
+
 
 @api_view(['GET'])
 def get_relationship(request, slug):
@@ -132,12 +157,14 @@ def get_relationship(request, slug):
                 return Response({'relationship': 'invite-received'})
         return Response({'relationship': 'none'})
 
+
 @api_view(['GET'])
 def friend_requests_received(request):
     profile = Profile.objects.get(user=request.user)
     invites = Relationship.objects.invitations_received(profile)
     serializer = RelationshipSerializer(invites, many=True)
     return Response(serializer.data)
+
 
 @api_view(['GET'])
 def pending_sent_friend_requests(request):
@@ -146,15 +173,17 @@ def pending_sent_friend_requests(request):
     serializer = RelationshipSerializer(relationships, many=True)
     return Response(serializer.data)
 
+
 @api_view(['POST'])
 def remove_from_friends(request):
     if request.method == 'POST':
         profile = Profile.objects.get(user=request.user)
         other_profile = Profile.objects.get(id=int(request.data))
-        relationship = Relationship.objects.filter(Q(sender=profile) | Q(sender=other_profile), Q(receiver=profile) | Q(receiver=other_profile), status='accepted').first()
+        relationship = Relationship.objects.filter(Q(sender=profile) | Q(sender=other_profile), Q(
+            receiver=profile) | Q(receiver=other_profile), status='accepted').first()
         relationship.delete()
         return Response('Removed from friends with success')
-        
+
 
 @api_view(['POST'])
 def send_friends_request(request):
@@ -163,24 +192,29 @@ def send_friends_request(request):
         receiver = Profile.objects.get(id=int(request.data))
         if Relationship.objects.filter(Q(sender=sender) | Q(sender=receiver), Q(receiver=sender) | Q(receiver=receiver)).exists():
             return Response('Users already have a relationship')
-        Relationship.objects.create(sender=sender, receiver=receiver, status='sent')
+        Relationship.objects.create(
+            sender=sender, receiver=receiver, status='sent')
         return Response('Friend request sent')
+
 
 @api_view(['POST'])
 def cancel_friend_request(request):
     if request.method == 'POST':
         sender = Profile.objects.get(user=request.user)
         receiver = Profile.objects.get(id=int(request.data))
-        r = Relationship.objects.get(sender=sender, receiver=receiver, status='sent')
+        r = Relationship.objects.get(
+            sender=sender, receiver=receiver, status='sent')
         r.delete()
         return Response('Friend request canceled')
+
 
 @api_view(['POST'])
 def reply_friend_request(request):
     profile = Profile.objects.get(user=request.user)
     sender = Profile.objects.get(id=int(request.data['senderid']))
     reply = request.data['reply']
-    r = Relationship.objects.get(sender=sender, receiver=profile, status='sent')
+    r = Relationship.objects.get(
+        sender=sender, receiver=profile, status='sent')
     if reply == 'accept':
         r.status = 'accepted'
         r.save()
@@ -188,15 +222,20 @@ def reply_friend_request(request):
         r.delete()
     return Response('Replied with success')
 
+
 @api_view(['POST'])
 def block_profile(request):
     profile = Profile.objects.get(user=request.user)
     profile_to_block = Profile.objects.get(id=request.data['id'])
-    for r in Relationship.objects.filter(Q(sender=profile) | Q(sender=profile_to_block), Q(receiver=profile) | Q(receiver=profile_to_block)):
-        r.delete()
-    profile.blocked_users.add(profile_to_block.user)
-    profile.save()
-    return Response('Profile blocked')
+    if profile != profile_to_block:
+        for r in Relationship.objects.filter(Q(sender=profile) | Q(sender=profile_to_block), Q(receiver=profile) | Q(receiver=profile_to_block)):
+            r.delete()
+            profile.blocked_users.add(profile_to_block.user)
+            profile.save()
+        return Response('Profile blocked')
+    else:
+        return Response('You can\'t block yourself')
+
 
 @api_view(['POST'])
 def unblock_profile(request):
@@ -206,15 +245,18 @@ def unblock_profile(request):
     profile.save()
     return Response('Profile unblocked')
 
+
 @api_view(['POST'])
 def set_myinterests(request):
     profile = Profile.objects.get(user=request.user)
     profile.interests.clear()
     for title in request.data['public_interests']:
-        public_i, created = Interest.objects.get_or_create(title=title.lower(), public=True)
+        public_i, created = Interest.objects.get_or_create(
+            title=title.lower(), public=True)
         profile.interests.add(public_i)
     for title in request.data['private_interests']:
-        private_i, created = Interest.objects.get_or_create(title=title.lower(), public=False)
+        private_i, created = Interest.objects.get_or_create(
+            title=title.lower(), public=False)
         profile.interests.add(private_i)
     profile.save()
     return Response('Interests updated')
